requirements.txt
pytelegrambotapi==4.12.0
python-dotenv==0.21.0
requests
peewee
python-telegram-bot-calendar